+++ src/runner.rs
@@ -2,12 +2,16 @@ use std::{env, io::Read, path::PathBuf, str::FromStr};
 
 use crate::{
     devrcfile::Devrcfile,
+    docs::DocHelper,
     errors::{DevrcError, DevrcResult},
     interrupt::setup_interrupt_handler,
     raw_devrcfile::RawDevrcfile,
     scope::Scope,
     utils,
-    utils::{get_absolute_path, get_global_devrc_file, get_local_devrc_file},
+    utils::{
+        get_absolute_path, get_directory_devrc_file, get_global_devrc_file,
+        get_local_user_defined_devrc_file,
+    },
 };
 
 use std::fmt::Debug;
@@ -23,6 +27,10 @@ pub struct Runner {
 
     /// Assembled tasks library
     pub devrc: Devrcfile,
+
+    pub global_loaded: bool,
+
+    pub loaded_files: Vec<PathBuf>,
 }
 
 impl Runner {
@@ -35,15 +43,21 @@ impl Runner {
             rest: vec![],
             dry_run: false,
             devrc: Devrcfile::default(),
+            global_loaded: false,
+            loaded_files: vec![],
         }
     }
 
-    pub fn dry_run(&mut self) {
-        self.dry_run = true;
+    pub fn setup_dry_run(&mut self, dry_run: bool) {
+        self.dry_run = dry_run;
     }
 
-    pub fn get_global_rawdevrc_file(&self) -> DevrcResult<RawDevrcfile> {
-        if let Some(value) = get_global_devrc_file() {
+    // Try to get devrcfile
+    pub fn get_rawdevrc_file<F>(&self, try_file_func: F) -> DevrcResult<RawDevrcfile>
+    where
+        F: Fn() -> Option<PathBuf>,
+    {
+        if let Some(value) = try_file_func() {
             match RawDevrcfile::from_file(&value) {
                 Ok(mut parsed_file) => {
                     parsed_file.setup_path(value)?;
@@ -52,43 +66,74 @@ impl Runner {
                 Err(error) => return Err(error),
             }
         }
-        Err(DevrcError::GlobalNotExists)
+        Err(DevrcError::NotExists)
     }
 
-    pub fn get_local_rawdevrc_file(&self) -> DevrcResult<RawDevrcfile> {
-        if let Some(value) = get_local_devrc_file() {
-            match RawDevrcfile::from_file(&value) {
-                Ok(mut parsed_file) => {
-                    parsed_file.setup_path(value)?;
-                    return Ok(parsed_file);
-                }
-                Err(error) => return Err(error),
+    pub fn load_global(&mut self) -> DevrcResult<()> {
+        if let Ok(devrcfile) = self.get_rawdevrc_file(get_global_devrc_file) {
+            if let Some(path) = devrcfile.path.clone() {
+                self.add_loaded_file(path)?;
             }
+            self.devrc.add_raw_devrcfile(devrcfile)?;
         }
-        Err(DevrcError::LocalNotExists)
+        self.global_loaded = true;
+
+        Ok(())
+    }
+
+    pub fn load_file<F>(&mut self, try_file_func: F) -> DevrcResult<()>
+    where
+        F: Fn() -> Option<PathBuf>,
+    {
+        if let Ok(devrcfile) = self.get_rawdevrc_file(try_file_func) {
+            // Load global file if option is enabled before adding current file
+            if devrcfile.is_global_enabled() && self.global_loaded {
+                self.load_global()?
+            }
+            if let Some(path) = devrcfile.path.clone() {
+                self.add_loaded_file(path)?;
+            }
+            self.devrc.add_raw_devrcfile(devrcfile)?;
+        }
+
+        Ok(())
     }
 
     pub fn load(&mut self) -> DevrcResult<()> {
-        if let Ok(devrcfile) = self.get_global_rawdevrc_file() {
-            self.devrc.add_raw_devrcfile(devrcfile)?;
+        // Try to load global devrcfile if flag is enabled
+        if self.use_global {
+            self.load_global()?;
         }
 
-        for file in self.files.iter() {
-            match RawDevrcfile::from_file(file) {
-                Ok(parsed_file) => {
-                    let mut parsed_file: RawDevrcfile = parsed_file;
-                    parsed_file.setup_path(file.to_path_buf())?;
-                    dbg!(&parsed_file);
-                    self.devrc.add_raw_devrcfile(parsed_file)?;
+        // Load files only if option specified
+        if !self.files.is_empty() {
+            let mut loaded_files: Vec<PathBuf> = Vec::new();
+
+            for file in &self.files {
+                match self.get_rawdevrc_file(|| Some(file.to_path_buf())) {
+                    Ok(devrcfile) => {
+                        if let Some(path) = devrcfile.path.clone() {
+                            loaded_files.push(path);
+                        }
+                        self.devrc.add_raw_devrcfile(devrcfile)?;
+                    }
+                    Err(error) => return Err(error),
                 }
-                Err(error) => return Err(error),
-            };
+            }
+
+            for file in loaded_files {
+                self.add_loaded_file(file)?;
+            }
+        } else {
+            // Try to load Devrcfile from current directory
+
+            self.load_file(get_directory_devrc_file)?;
+
+            // Try to load Devrcfile.local
+            self.load_file(get_local_user_defined_devrc_file)?;
         }
 
-        if let Ok(devrcfile) = self.get_local_rawdevrc_file() {
-            self.devrc.add_raw_devrcfile(devrcfile)?;
-        }
-
+        self.devrc.setup_dry_run(self.dry_run)?;
         Ok(())
     }
 
@@ -104,6 +149,12 @@ impl Runner {
             }
             Err(error) => return Err(error),
         }
+        self.devrc.setup_dry_run(self.dry_run)?;
+        Ok(())
+    }
+
+    pub fn add_loaded_file(&mut self, file: PathBuf) -> DevrcResult<()> {
+        self.loaded_files.push(file);
         Ok(())
     }
 
@@ -134,10 +185,8 @@ impl Runner {
     /// Show tasks list with short descriptions
     pub fn list_tasks(&self) -> DevrcResult<()> {
         println!("Available devrc tasks:");
-
-        // TODO: remove copy
         let (max_taskname_width, _) = self.devrc.get_max_taskname_width();
-        for (name, task) in self.devrc.tasks.items.clone() {
+        for (name, task) in self.devrc.get_tasks_docs() {
             let help = task.format_help()?;
 
             if name.starts_with('_') {
@@ -158,26 +207,65 @@ impl Runner {
         Ok(())
     }
 
+    /// Show global variables and their computed values
     pub fn list_vars(&self) -> DevrcResult<()> {
-        println!("List devrc variables:");
+        println!("List global devrc variables:");
+
+        let max_variable_name_width = self.devrc.variables.get_max_key_width();
+
+        for (name, value) in self.devrc.get_vars() {
+            println!(
+                "{:width$}{:max_variable_name_width$} = {}",
+                "",
+                name,
+                value,
+                width = 2,
+                max_variable_name_width = max_variable_name_width
+            );
+        }
+        Ok(())
+    }
+
+    /// Show global environment variables and their computed values
+    pub fn list_env_vars(&self) -> DevrcResult<()> {
+        println!("List global devrc environment variables:");
+
+        let max_variable_name_width = self.devrc.environment.get_max_key_width();
+
+        for (name, value) in self.devrc.get_environment_vars() {
+            println!(
+                "{:width$}{:max_variable_name_width$} = {}",
+                "",
+                name,
+                value,
+                width = 2,
+                max_variable_name_width = max_variable_name_width
+            );
+        }
         Ok(())
     }
 
     /// Load global devrc
     pub fn use_global(&mut self) {
-        println!("Use global devrc file");
         self.use_global = true;
     }
 
-    //  else if utils::is_global_devrc_file_exists() {
-    //     config.add_file(utils::get_global_devrc_file().unwrap());
-    // } else {
-    //     println!("Show help")
-    // }
-
     /// Show description for given task, variable or environment variable
-    pub fn describe(&self, _params: Vec<String>) -> DevrcResult<()> {
-        println!("Describe task or variable");
+    pub fn describe(&self, params: Vec<String>) -> DevrcResult<()> {
+        for name in params {
+            let task = self.devrc.find_task(&name)?;
+            println!(
+                "Usage: {}\nDescription: {}",
+                task.get_usage_help(&name)?,
+                task.format_help()?,
+                // width = 2,
+            );
+
+            if let Some(example) = task.get_example() {
+                println!("Examples: \n{}", example);
+            }
+            println!();
+        }
         Ok(())
     }
 
@@ -188,30 +276,38 @@ impl Runner {
 
         self.rest = params;
 
+        info!(
+            "Global defined interpreter: `{:}`",
+            &self.devrc.config.interpreter
+        );
+
         if let Some(value) = get_global_devrc_file() {
-            info!("Global devrcfile exists: {:?}", value);
+            info!("Global .devrc exists: {:?}", value);
         }
 
-        if let Some(value) = get_local_devrc_file() {
-            info!("Local devrcfile exists: {:?}", value);
+        if let Some(value) = get_directory_devrc_file() {
+            info!("Local directory Devrcfile exists: {:?}", value);
+        }
+
+        if let Some(value) = get_local_user_defined_devrc_file() {
+            info!("Local user defined Devrcfile.local exists: {:?}", value);
         }
 
         for file in &self.files {
             if let Ok(file) = get_absolute_path(&file, None) {
                 if file.exists() {
-                    info!("Given devrcfile exists: {:?}", file);
+                    info!("Given Devrcfile exists: {:?}", file);
                 } else {
-                    info!("Given devrcfile not exists: {:?}", file);
+                    info!("Given Devrcfile not exists: {:?}", file);
                 }
             } else {
-                error!("Given devrcfile with broken path: {:?}", &file);
+                error!("Given Devrcfile with broken path: {:?}", &file);
             }
         }
 
-        info!(
-            "Global defined interpreter: `{:}`",
-            &self.devrc.config.interpreter
-        );
+        for file in &self.loaded_files {
+            info!("Loaded file: {:?}", file);
+        }
 
         dbg!(self);
     }